+++ src/js/core/video.js
@@ -1,5 +1,5 @@
-import { hasAttr, isInView, isTag, isVideo, isVisible, mute, pause, play } from 'uikit-util';
-import { intersection, resize } from '../api/observables';
+import { hasAttr, isTag, isVideo, mute, parent, pause, play } from 'uikit-util';
+import { intersection } from '../api/observables';
 
 export default {
     args: 'autoplay',
@@ -15,9 +15,7 @@ export default {
     },
 
     connected() {
-        this.inView = this.autoplay === 'inview';
-
-        if (this.inView && !hasAttr(this.$el, 'preload')) {
+        if (this.autoplay === 'inview' && !hasAttr(this.$el, 'preload')) {
             this.$el.preload = 'none';
         }
 
@@ -30,29 +28,18 @@ export default {
         }
     },
 
-    observe: [intersection({ args: { intersecting: false } }), resize()],
-
-    update: {
-        read({ visible }) {
-            if (!isVideo(this.$el)) {
-                return false;
-            }
-
-            return {
-                prev: visible,
-                visible: isVisible(this.$el),
-                inView: this.inView && isInView(this.$el),
-            };
-        },
-
-        write({ prev, visible, inView }) {
-            if (!visible || (this.inView && !inView)) {
-                pause(this.$el);
-            } else if ((this.autoplay === true && !prev) || inView) {
-                play(this.$el);
-            }
-        },
-
-        events: ['resize'],
-    },
+    observe: [
+        intersection({
+            filter: ({ $el }) => isVideo($el),
+            handler([{ isIntersecting }]) {
+                if (isIntersecting) {
+                    play(this.$el);
+                } else {
+                    pause(this.$el);
+                }
+            },
+            args: { intersecting: false },
+            options: ({ $el, autoplay }) => ({ root: autoplay === 'inview' ? null : parent($el) }),
+        }),
+    ],
 };